move fs.go to a new internal.fs package
Avoids the anti-pattern of using a 'util' package.
