+++ xfs/xfs.go
@@ -19,7 +19,7 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/prometheus/procfs/internal/util"
+	"github.com/prometheus/procfs/internal/fs"
 )
 
 // Stats contains XFS filesystem runtime statistics, parsed from
@@ -173,8 +173,8 @@ type ExtendedPrecisionStats struct {
 // FS represents the pseudo-filesystems proc and sys, which provides an interface to
 // kernel data structures.
 type FS struct {
-	proc *util.FS
-	sys  *util.FS
+	proc *fs.FS
+	sys  *fs.FS
 }
 
 // DefaultProcMountPoint is the common mount point of the proc filesystem.
@@ -189,14 +189,14 @@ func NewFS(procMountPoint string, sysMountPoint string) (FS, error) {
 	if strings.TrimSpace(procMountPoint) == "" {
 		procMountPoint = DefaultProcMountPoint
 	}
-	procfs, err := util.NewFS(procMountPoint)
+	procfs, err := fs.NewFS(procMountPoint)
 	if err != nil {
 		return FS{}, err
 	}
 	if strings.TrimSpace(sysMountPoint) == "" {
 		sysMountPoint = DefaultSysMountPoint
 	}
-	sysfs, err := util.NewFS(sysMountPoint)
+	sysfs, err := fs.NewFS(sysMountPoint)
 	if err != nil {
 		return FS{}, err
 	}
